Fix assembler error messsage panic
Under specific circumstances, an error message may be generated for
the final characters of the source code line, in which case the
`nom_locate` column may not represent the position of the error in
the string, since `nom` counts `\n`.

With this fix, the position of the error is the minimum between the
`nom_locate` position and the length of the source string, this
eliminating the panic in all circumstances.